requirements.txt
flowserv-core>=0.4.1
flask
flask_cors
